refactor(app); move some liquid stuff to xform
