+++ app/src/organisms/ODD/ProtocolSetup/ProtocolSetupLiquids/index.tsx
@@ -23,7 +23,7 @@ import { ODDBackButton } from '/app/molecules/ODDBackButton'
 import { SmallButton } from '/app/atoms/buttons'
 
 import { useMostRecentCompletedAnalysis } from '../../../LabwarePositionCheck/useMostRecentCompletedAnalysis'
-import { getTotalVolumePerLiquidId } from '../../../Devices/ProtocolRun/SetupLiquids/utils'
+import { getTotalVolumePerLiquidId } from '/app/transformations/analysis'
 import { LiquidDetails } from './LiquidDetails'
 import type { ParsedLiquid, RunTimeCommand } from '@opentrons/shared-data'
 import type { SetupScreens } from '../types'
