+++ app/src/transformations/analysis/index.ts
@@ -2,3 +2,4 @@ export * from './parseProtocolAnalysisOutput'
 export * from './getProtocolModulesInfo'
 export * from './getLabwareOffsetLocation'
 export * from './getLabwareRenderInfo'
+export * from './liquids'
